added comments, moved helper class to separate file
