+++ src/Artemis.Core/Services/ColorQuantizer/ColorType.cs
@@ -1,0 +1,12 @@
+﻿namespace Artemis.Core.Services
+{
+    public enum ColorType
+    {
+        Vibrant,
+        LightVibrant,
+        DarkVibrant,
+        Muted,
+        LightMuted,
+        DarkMuted
+    }
+}
